backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
google-generativeai==0.5.4
openai==1.30.1

python-dotenv==1.0.1
websockets==12.0
aiosqlite==0.20.0
prisma==0.13.1
groq==1.1.1
